feat(healthcheck): support native lua healthchecks, and fixes
After neovim/neovim#15259 lua healthchecks are called directly and are
prefered over neovim autoload ones. In the case of nvim-treesitter, the
discrepancy between the autoload function nvim_treesitter and the lua
base module nvim-treesitter makes it neccessary to add a custom login in
a guard.

Also the mentioned PR changed the design of healthchecks to not use
output capturing, therefore avoid printing and instead concatenate and
call the health functions.